/ui: renamed `ScreenGauge` component to `ScreenSensor`
Closes #354
