+++ packages/components/ui/src/sensors/index.js
@@ -1,0 +1,1 @@
+export {default as ScreenSensor} from './screen/ScreenSensor.svelte';
