+++ packages/components/ui/src/sensors/screen/ScreenSensor.svelte
